requirements.txt
Django==1.10.2
djangorestframework==3.5.1
httpie==0.9.6
Pygments==2.1.3
requests==2.11.1
